refactor: remove dead vault modules
